split out websocket api and webserver
