+++ music_assistant/server/providers/dlna/helpers.py
@@ -23,7 +23,7 @@ class DLNANotifyServer(UpnpNotifyServer):
         """Initialize."""
         self.mass = mass
         self.event_handler = UpnpEventHandler(self, requester)
-        self.mass.webapp.router.add_route("NOTIFY", "/notify", self._handle_request)
+        self.mass.webserver.register_route("/notify", self._handle_request, method="NOTIFY")
 
     async def _handle_request(self, request: Request) -> Response:
         """Handle incoming requests."""
@@ -40,4 +40,4 @@ class DLNANotifyServer(UpnpNotifyServer):
     @property
     def callback_url(self) -> str:
         """Return callback URL on which we are callable."""
-        return f"{self.mass.base_url}/notify"
+        return f"{self.mass.webserver.base_url}/notify"
