--- music_assistant/server/providers/frontend/manifest.json
@@ -1,12 +1,0 @@
-{
-  "type": "plugin",
-  "domain": "frontend",
-  "name": "Frontend",
-  "description": "The default Music Assistant (web) frontend, written in Vue, hosted within the Music Assistant server.",
-  "codeowners": ["@music-assistant"],
-  "requirements": ["music-assistant-frontend==20230319.1"],
-  "documentation": "",
-  "multi_instance": false,
-  "builtin": true,
-  "load_by_default": true
-}
