+++ music_assistant/server/server.py
@@ -2,18 +2,16 @@
 from __future__ import annotations
 
 import asyncio
-import importlib
-import inspect
 import logging
 import os
 from collections.abc import Awaitable, Callable, Coroutine
 from typing import TYPE_CHECKING, Any
 from uuid import uuid4
 
-from aiohttp import ClientSession, TCPConnector, web
+from aiohttp import ClientSession, TCPConnector
 from zeroconf import InterfaceChoice, NonUniqueNameException, ServiceInfo, Zeroconf
 
-from music_assistant.common.helpers.util import get_ip, get_ip_pton, select_free_port
+from music_assistant.common.helpers.util import get_ip, get_ip_pton
 from music_assistant.common.models.config_entries import ProviderConfig
 from music_assistant.common.models.enums import EventType, ProviderType
 from music_assistant.common.models.errors import SetupFailedError
@@ -26,18 +24,15 @@ from music_assistant.server.controllers.metadata import MetaDataController
 from music_assistant.server.controllers.music import MusicController
 from music_assistant.server.controllers.players import PlayerController
 from music_assistant.server.controllers.streams import StreamsController
-from music_assistant.server.helpers.api import APICommandHandler, api_command, mount_websocket_api
-from music_assistant.server.helpers.util import install_package
-from music_assistant.server.models.plugin import PluginProvider
+from music_assistant.server.controllers.webserver import WebserverController
+from music_assistant.server.helpers.api import APICommandHandler, api_command
+from music_assistant.server.helpers.util import get_provider_module
 
-from .models.metadata_provider import MetadataProvider
-from .models.music_provider import MusicProvider
-from .models.player_provider import PlayerProvider
+from .models import ProviderInstanceType
 
 if TYPE_CHECKING:
     from types import TracebackType
 
-ProviderInstanceType = MetadataProvider | MusicProvider | PlayerProvider
 EventCallBackType = Callable[[MassEvent], None]
 EventSubscriptionType = tuple[EventCallBackType, tuple[EventType] | None, tuple[str] | None]
 
@@ -52,25 +47,21 @@ class MusicAssistant:
 
     loop: asyncio.AbstractEventLoop
     http_session: ClientSession
-    _web_apprunner: web.AppRunner
-    _web_tcp: web.TCPSite
 
-    def __init__(self, storage_path: str, port: int | None = None) -> None:
+    def __init__(self, storage_path: str) -> None:
         """Initialize the MusicAssistant Server."""
         self.storage_path = storage_path
-        self.port = port
         self.base_ip = get_ip()
         # shared zeroconf instance
         self.zeroconf = Zeroconf(interfaces=InterfaceChoice.All)
-        # we dynamically register command handlers
-        self.webapp = web.Application()
+        # we dynamically register command handlers which can be consumed by the apis
         self.command_handlers: dict[str, APICommandHandler] = {}
         self._subscribers: set[EventSubscriptionType] = set()
         self._available_providers: dict[str, ProviderManifest] = {}
         self._providers: dict[str, ProviderInstanceType] = {}
-
         # init core controllers
         self.config = ConfigController(self)
+        self.webserver = WebserverController(self)
         self.cache = CacheController(self)
         self.metadata = MetaDataController(self)
         self.music = MusicController(self)
@@ -84,7 +75,6 @@ class MusicAssistant:
     async def start(self) -> None:
         """Start running the Music Assistant server."""
         self.loop = asyncio.get_running_loop()
-
         # create shared aiohttp ClientSession
         self.http_session = ClientSession(
             loop=self.loop,
@@ -92,35 +82,21 @@ class MusicAssistant:
         )
         # setup config controller first and fetch important config values
         await self.config.setup()
-        if self.port is None:
-            # if port is None, we need to autoselect it
-            self.port = await select_free_port(8095, 9200)
-        # allow overriding of the base_ip if autodetect failed
         self.base_ip = self.config.get(CONF_WEB_IP, self.base_ip)
         LOGGER.info(
-            "Starting Music Assistant Server (%s) on port: %s - autodetected IP-address: %s",
+            "Starting Music Assistant Server (%s) - autodetected IP-address: %s",
             self.server_id,
-            self.port,
             self.base_ip,
         )
-
         # setup other core controllers
         await self.cache.setup()
+        await self.webserver.setup()
         await self.music.setup()
         await self.metadata.setup()
         await self.players.setup()
         await self.streams.setup()
-
         # load providers
         await self._load_providers()
-        # setup web server
-        mount_websocket_api(self, "/ws")
-        self._web_apprunner = web.AppRunner(self.webapp, access_log=None)
-        await self._web_apprunner.setup()
-        # set host to None to bind to all addresses on both IPv4 and IPv6
-        host = None
-        self._web_tcp = web.TCPSite(self._web_apprunner, host=host, port=self.port)
-        await self._web_tcp.start()
         self._setup_discovery()
 
     async def stop(self) -> None:
@@ -131,31 +107,21 @@ class MusicAssistant:
         # cancel all running tasks
         for task in self._tracked_tasks.values():
             task.cancel()
-        # stop/clean streams controller
-        await self.streams.close()
-        # stop/clean webserver
-        await self._web_tcp.stop()
-        await self._web_apprunner.cleanup()
-        await self.webapp.shutdown()
-        await self.webapp.cleanup()
+        # cleanup all providers
+        for prov_id in list(self._providers.keys()):
+            await self.unload_provider(prov_id)
         # stop core controllers
+        await self.streams.close()
         await self.metadata.close()
         await self.music.close()
         await self.players.close()
-        # cleanup all providers
-        for prov in self._providers.values():
-            await prov.close()
+        await self.webserver.close()
         # cleanup cache and config
         await self.config.close()
         await self.cache.close()
         # close/cleanup shared http session
         if self.http_session:
             await self.http_session.close()
-
-    @property
-    def base_url(self) -> str:
-        """Return the (web)server's base url."""
-        return f"http://{self.base_ip}:{self.port}"
 
     @property
     def server_id(self) -> str:
@@ -191,9 +157,11 @@ class MusicAssistant:
         if prov is not None and (return_unavailable or prov.available):
             return prov
         for prov in self._providers.values():
-            if prov.domain == provider_instance_or_domain and return_unavailable or prov.available:
+            if prov.domain != provider_instance_or_domain:
+                continue
+            if return_unavailable or prov.available:
                 return prov
-        LOGGER.warning("Provider {provider_instance_or_domain} is not available")
+        LOGGER.debug("Provider %s is not available", provider_instance_or_domain)
         return None
 
     def signal_event(
@@ -337,29 +305,9 @@ class MusicAssistant:
                 )
 
         # try to load the module
-        prov_mod = importlib.import_module(f".{domain}", "music_assistant.server.providers")
-        for name, obj in inspect.getmembers(prov_mod):
-            if not inspect.isclass(obj):
-                continue
-            # lookup class to initialize
-            if name == prov_manifest.init_class or (
-                not prov_manifest.init_class
-                and issubclass(
-                    obj, MusicProvider | PlayerProvider | MetadataProvider | PluginProvider
-                )
-                and obj != MusicProvider
-                and obj != PlayerProvider
-                and obj != MetadataProvider
-                and obj != PluginProvider
-            ):
-                prov_cls = obj
-                break
-        else:
-            raise AttributeError("Unable to locate Provider class")
-        provider: ProviderInstanceType = prov_cls(self, prov_manifest, conf)
-
+        prov_mod = await get_provider_module(domain)
         try:
-            await asyncio.wait_for(provider.setup(), 30)
+            provider = await asyncio.wait_for(prov_mod.setup(self, prov_manifest, conf), 30)
         except TimeoutError as err:
             raise SetupFailedError(f"Provider {domain} did not load within 30 seconds") from err
         # if we reach this point, the provider loaded successfully
@@ -384,7 +332,7 @@ class MusicAssistant:
                 if sync_task.provider_instance == instance_id:
                     sync_task.task.cancel()
                     await sync_task.task
-            await provider.close()
+            await provider.unload()
             self._providers.pop(instance_id)
             self.signal_event(EventType.PROVIDERS_UPDATED, data=self.get_providers())
 
@@ -415,7 +363,7 @@ class MusicAssistant:
         for prov_manifest in self._available_providers.values():
             if not prov_manifest.load_by_default:
                 continue
-            self.config.create_default_provider_config(prov_manifest.domain)
+            await self.config.create_default_provider_config(prov_manifest.domain)
 
         async def load_provider(prov_conf: ProviderConfig) -> None:
             """Try to load a provider and catch errors."""
@@ -434,8 +382,9 @@ class MusicAssistant:
                 self.config.set(f"{CONF_PROVIDERS}/{prov_conf.instance_id}/last_error", str(exc))
 
         # load all configured (and enabled) providers
+        prov_configs = await self.config.get_provider_configs()
         async with asyncio.TaskGroup() as tg:
-            for prov_conf in self.config.get_provider_configs():
+            for prov_conf in prov_configs:
                 if not prov_conf.enabled:
                     continue
                 tg.create_task(load_provider(prov_conf))
@@ -456,9 +405,6 @@ class MusicAssistant:
                 try:
                     provider_manifest = await ProviderManifest.parse(file_path)
                     self._available_providers[provider_manifest.domain] = provider_manifest
-                    # install requirement/dependencies
-                    for requirement in provider_manifest.requirements:
-                        await install_package(requirement)
                     LOGGER.debug("Loaded manifest for provider %s", dir_str)
                 except Exception as exc:  # pylint: disable=broad-except
                     LOGGER.exception(
@@ -476,7 +422,7 @@ class MusicAssistant:
             zeroconf_type,
             name=f"{server_id}.{zeroconf_type}",
             addresses=[get_ip_pton()],
-            port=self.port,
+            port=self.webserver.port,
             properties={},
             server=f"mass_{server_id}.local.",
         )
