+++ pyproject.toml
@@ -34,6 +34,7 @@ server = [
   "python-slugify==8.0.1",
   "mashumaro==3.5.0",
   "memory-tempfile==2.2.3",
+  "music-assistant-frontend==20230319.1",
   "pillow==9.4.0",
   "unidecode==1.3.6",
   "xmltodict==0.13.0",
